Update appearance and function of suites
The website is a essentially a portfolio relying heavily on images.
Created functional components that takes images and title of dance
as props and displays upon them in a parallax banner.

A suite overview component is made to showcase all the suite groups
that are in the show.

This commit is creates the components needed to showcase 1 per
suite.
Simple dummy text is in place of the summary until the summary
prop is created.