Implemented Directors Page and Clean Up
Created a directors page containing images and names of all the pcn
directors and their names. Lorem Ipsum text is in place of a letter
or thoughts to be posted

The HomePage component contains 5 pages now instead of 4. All the
pictures are in the right places. The parallax effect is cleaned
up for consistency.

Still need access to facts, trivia, and information to add to the
site. Need to come up with a mobile version of the website in
general and have graphics that are more mobile friendly

Idea: sponsors?