Implemented Footer and Layout Changes
Created a stateless component as the footer of the website. Contains
message saying who designed the site and has icons with links to
facebook, twitter, and Instagram.

Layout changes have been made. Different styled font on the AboutPage
that highlights some trivia and contains picture of the full cast.

Added fullCast photo to the assets of info folder.