Clean up home page and incorporate changes
Implemented styled components for the main page. Inline css was getting
out of hand and messy so styled components was created and imported from
Primitives file.

Incorporated changes to Events and font changes.

Removed navigation bar. It only exists in PCN page. The homepage is now
a large logo in the center with 2 arrows to scroll down for more
information or click right to go to PCN page.