Implemented styled components and clean up styles
Inline CSS was utilized everywhere and often repetitive. Created styled
components to modularize each part of the page. Page is a styled div
along with row, column, span, title, and tag.

Created Primitives file for home and pcn to use separately.

Got rid of all the side margins for PCN. All navigation will be
available on the first section of each page with arrows to indicate
which way to go.

EventDetails has a back button to exit.

Footer has been modified to all the social media accounts are on top
while site credit is on the bottom.

The entire site is now configured for smartphones. Have yet to test for
tablets.