Implemented multiple pages and design changes
Added pages for Friendship Games, Barrio Fiesta, and EOTYD. They all
have the imagefade gallery with their respective photos in the assets
folder and exported through a data.js file.

Home Page now has a different background that is a blue which changes
from dark to light.

ImageFade has been modified to take in an interval as a prop to go
through each of the photos.

PCN home page has a major design change which is replicated from the KAA
page. Gallery to be added soon.

Friendship Games, Barrio, and EOTYD all need information about them to
be added in the next update.